# Defer formatting in traces to make them cheaper