--- a/fdbclient/FDBTypes.h
+++ b/fdbclient/FDBTypes.h
@@ -503,6 +503,13 @@ struct RangeResultRef : VectorRef<KeyValueRef> {
 	}
 };
 
+template<>
+struct Traceable<RangeResultRef> : std::true_type {
+	static std::string toString(const RangeResultRef& value) {
+		return Traceable<VectorRef<KeyValueRef>>::toString(value);
+	}
+};
+
 struct KeyValueStoreType {
 	// These enumerated values are stored in the database configuration, so can NEVER be changed.  Only add new ones just before END.
 	enum StoreType {