@mpilman 2019-03-15
It seems that this currently only compiles with clang - will look into this tomorrow.

However, I ran some first primitive test (the test can be found in `ActorCollection.actor.cpp`). These are the numbers I am getting:

|       | Without optimization | With optimization | Speedup |
| --- | --------------------- | ----------------- | --------- |
| 10.000.000 Traces to file | 4.1 s | 1.7 s | 241% |
| 100.000.000 Disabled debug traces | 5.4 s | 4.5 s | 20% |

I think these numbers can be explained like this:
- For disabled traces, we mostly save on inlining the detail function and if it is disabled, we save an actual function call
- For enabled traces, this solution might still get rid of some allocations. I would have expected to save on those in for disabled traces as well, but it might be that the compiler does make some fancy optimizations.