@mpilman 2019-03-15
I added a `printable`-trace to both of those in the microbenchmark (one uses `printable`, this one uses the templated mechanism). Here are the results on my Linux-box:

|       | Without optimization | With optimization | Speedup |
| --- | --------------------- | ----------------- | --------- |
| 10.000.000 Traces to file | 8 s | 4.7 s | 70% |
| 100.000.000 Disabled debug traces | 47 s | 16.5 s | 285% |