@mpilman 2019-03-15
I did some last testing after improving the micro-benchmark a bit (I mostly pre-generate all random numbers to only measure the actual trace-calls). It is pretty clear that this new solution is significantly faster:

|       | Without optimization | With optimization | Speedup |
| --- | --------------------- | ----------------- | --------- |
| 10.000.000 Traces to file | 8.1 s | 4.4 s | 84% |
| 100.000.000 Disabled debug traces | 52.3 s | 11.1 s | 471% |